brandradar-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
textblob==0.17.1
praw==7.7.1
python-dotenv==1.0.0